chore(Poll): move PollVotersDetails, add constants
